[이범용] 버츄얼 패드 연동 관련 Movement 로직 수정  - FSM 상태관련 처리 로직 수정

+++ Source/GameService/GameObject/State/GsStateBase.cpp
@@ -17,7 +17,7 @@ IGsStateBase::~IGsStateBase()
 	return !IsSameState(StateID);
 }*/
 
-bool IGsStateBase::IsSameState(int StateID)
+bool IGsStateBase::IsSameState(uint8 StateID)
 {
 	return (GetStateID() == StateID);
 }
